Split nodedef and general locale into 2 actions
