+++ scripts/collect-i18n-general.ts
@@ -1,19 +1,13 @@
 import * as fs from 'fs'
 
-import {
-  ComfyPage,
-  comfyPageFixture as test
-} from '../browser_tests/fixtures/ComfyPage'
+import { comfyPageFixture as test } from '../browser_tests/fixtures/ComfyPage'
 import { CORE_MENU_COMMANDS } from '../src/constants/coreMenuCommands'
 import { SERVER_CONFIG_ITEMS } from '../src/constants/serverConfig'
-import type { ComfyApi } from '../src/scripts/api'
 import type { ComfyCommandImpl } from '../src/stores/commandStore'
-import { ComfyNodeDefImpl } from '../src/stores/nodeDefStore'
 import type { FormItem, SettingParams } from '../src/types/settingTypes'
 import { formatCamelCase, normalizeI18nKey } from '../src/utils/formatUtil'
 
 const localePath = './src/locales/en/main.json'
-const nodeDefsPath = './src/locales/en/nodeDefs.json'
 const commandsPath = './src/locales/en/commands.json'
 const settingsPath = './src/locales/en/settings.json'
 
@@ -25,119 +19,7 @@ const extractMenuCommandLocaleStrings = (): Set<string> => {
   return labels
 }
 
-/**
- * Extracts the locales for node definitions, categories, and data types.
- *
- * @param comfyPage - The comfy page fixture.
- * @returns The locales for node definitions, categories, and data types.
- */
-const extractNodeDefsLocales = async (comfyPage: ComfyPage) => {
-  const nodeDefs: ComfyNodeDefImpl[] = Object.values(
-    await comfyPage.page.evaluate(async () => {
-      const api = window['app'].api as ComfyApi
-      return await api.getNodeDefs()
-    })
-  ).map((def) => new ComfyNodeDefImpl(def))
-
-  console.log(`Collected ${nodeDefs.length} node definitions`)
-
-  const allDataTypesLocale = Object.fromEntries(
-    nodeDefs
-      .flatMap((nodeDef) => {
-        const inputDataTypes = Object.values(nodeDef.inputs.all).map(
-          (inputSpec) => inputSpec.type
-        )
-        const outputDataTypes = nodeDef.outputs.all.map((output) => output.type)
-        const allDataTypes = [...inputDataTypes, ...outputDataTypes].flatMap(
-          (type: string) => type.split(',')
-        )
-        return allDataTypes.map((dataType) => [
-          normalizeI18nKey(dataType),
-          dataType
-        ])
-      })
-      .sort((a, b) => a[0].localeCompare(b[0]))
-  )
-
-  function extractInputs(nodeDef: ComfyNodeDefImpl) {
-    const inputs = Object.fromEntries(
-      nodeDef.inputs.all.flatMap((input) => {
-        const name = input.name
-        const tooltip = input.tooltip
-
-        if (name === undefined && tooltip === undefined) {
-          return []
-        }
-
-        return [
-          [
-            normalizeI18nKey(input.name),
-            {
-              name,
-              tooltip
-            }
-          ]
-        ]
-      })
-    )
-    return Object.keys(inputs).length > 0 ? inputs : undefined
-  }
-
-  function extractOutputs(nodeDef: ComfyNodeDefImpl) {
-    const outputs = Object.fromEntries(
-      nodeDef.outputs.all.flatMap((output, i) => {
-        // Ignore data types if they are already translated in allDataTypesLocale.
-        const name = output.name in allDataTypesLocale ? undefined : output.name
-        const tooltip = output.tooltip
-
-        if (name === undefined && tooltip === undefined) {
-          return []
-        }
-
-        return [
-          [
-            i.toString(),
-            {
-              name,
-              tooltip
-            }
-          ]
-        ]
-      })
-    )
-    return Object.keys(outputs).length > 0 ? outputs : undefined
-  }
-
-  const allNodeDefsLocale = Object.fromEntries(
-    nodeDefs
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .map((nodeDef) => [
-        normalizeI18nKey(nodeDef.name),
-        {
-          display_name: nodeDef.display_name ?? nodeDef.name,
-          description: nodeDef.description || undefined,
-          inputs: extractInputs(nodeDef),
-          outputs: extractOutputs(nodeDef)
-        }
-      ])
-  )
-
-  const allNodeCategoriesLocale = Object.fromEntries(
-    nodeDefs.flatMap((nodeDef) =>
-      nodeDef.category
-        .split('/')
-        .map((category) => [normalizeI18nKey(category), category])
-    )
-  )
-
-  return {
-    allNodeDefsLocale,
-    allNodeCategoriesLocale,
-    allDataTypesLocale
-  }
-}
-
-test('collect-i18n', async ({ comfyPage }) => {
+test('collect-i18n-general', async ({ comfyPage }) => {
   const commands = (
     await comfyPage.page.evaluate(() => {
       const workspace = window['app'].extensionManager
@@ -244,17 +126,6 @@ test('collect-i18n', async ({ comfyPage }) => {
     ])
   )
 
-  const COLLECT_NODE_DEFS = process.env.COLLECT_NODE_DEFS === 'true'
-
-  const { allNodeDefsLocale, allNodeCategoriesLocale, allDataTypesLocale } =
-    COLLECT_NODE_DEFS
-      ? await extractNodeDefsLocales(comfyPage)
-      : {
-          allNodeDefsLocale: undefined,
-          allNodeCategoriesLocale: undefined,
-          allDataTypesLocale: undefined
-        }
-
   fs.writeFileSync(
     localePath,
     JSON.stringify(
@@ -268,18 +139,13 @@ test('collect-i18n', async ({ comfyPage }) => {
           ...allSettingCategoriesLocale
         },
         serverConfigItems: allServerConfigsLocale,
-        serverConfigCategories: allServerConfigCategoriesLocale,
-        dataTypes: allDataTypesLocale ?? locale.dataTypes,
-        nodeCategories: allNodeCategoriesLocale ?? locale.nodeCategories
+        serverConfigCategories: allServerConfigCategoriesLocale
       },
       null,
       2
     )
   )
 
-  if (COLLECT_NODE_DEFS) {
-    fs.writeFileSync(nodeDefsPath, JSON.stringify(allNodeDefsLocale, null, 2))
-  }
   fs.writeFileSync(commandsPath, JSON.stringify(allCommandsLocale, null, 2))
   fs.writeFileSync(settingsPath, JSON.stringify(allSettingsLocale, null, 2))
 })
